[~] - package `io.qase.<framework>.module` was renamed to `io.qase.<framework>.guice.module` across the reporters
