+++ qase-cucumber3-jvm/src/main/java/io/qase/cucumber3/guice/module/Cucumber3Module.java
@@ -1,24 +1,24 @@
-package io.qase.cucumber5.module;
+package io.qase.cucumber3.guice.module;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
 import com.google.inject.Provides;
 import io.qase.api.config.apiclient.ApiClientConfigurer;
-import io.qase.cucumber5.configuration.Cucumber5ApiConfigurer;
+import io.qase.cucumber3.configuration.Cucumber3ApiConfigurer;
 import io.qase.guice.Injectors;
 import lombok.Getter;
 
-public class Cucumber5Module extends AbstractModule {
+public class Cucumber3Module extends AbstractModule {
 
     @Getter
     private static final Injector injector = initializeInjector();
 
     @Provides
     public ApiClientConfigurer apiClientConfigurer() {
-        return new Cucumber5ApiConfigurer();
+        return new Cucumber3ApiConfigurer();
     }
 
     private static Injector initializeInjector() {
-        return Injectors.createDefaultInjectorWithOverridingModules(new Cucumber5Module());
+        return Injectors.createDefaultInjectorWithOverridingModules(new Cucumber3Module());
     }
 }
