+++ qase-cucumber4-jvm/src/main/java/io/qase/cucumber4/QaseEventListener.java
@@ -11,7 +11,7 @@ import io.qase.api.utils.CucumberUtils;
 import io.qase.client.model.ResultCreate;
 import io.qase.client.model.ResultCreate.StatusEnum;
 import io.qase.client.model.ResultCreateSteps;
-import io.qase.cucumber4.module.Cucumber4Module;
+import io.qase.cucumber4.guice.module.Cucumber4Module;
 import lombok.AccessLevel;
 import lombok.Getter;
