+++ qase-cucumber5-jvm/src/main/java/io/qase/cucumber5/guice/module/Cucumber5Module.java
@@ -1,24 +1,24 @@
-package io.qase.junit5.module;
+package io.qase.cucumber5.guice.module;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
 import com.google.inject.Provides;
 import io.qase.api.config.apiclient.ApiClientConfigurer;
+import io.qase.cucumber5.configuration.Cucumber5ApiConfigurer;
 import io.qase.guice.Injectors;
-import io.qase.junit5.configuration.Junit5ApiConfigurer;
 import lombok.Getter;
 
-public class Junit5Module extends AbstractModule {
+public class Cucumber5Module extends AbstractModule {
 
     @Getter
     private static final Injector injector = initializeInjector();
 
     @Provides
     public ApiClientConfigurer apiClientConfigurer() {
-        return new Junit5ApiConfigurer();
+        return new Cucumber5ApiConfigurer();
     }
 
     private static Injector initializeInjector() {
-        return Injectors.createDefaultInjectorWithOverridingModules(new Junit5Module());
+        return Injectors.createDefaultInjectorWithOverridingModules(new Cucumber5Module());
     }
 }
