+++ qase-junit5/src/main/java/io/qase/junit5/guice/module/Junit5Module.java
@@ -1,24 +1,24 @@
-package io.qase.cucumber3.module;
+package io.qase.junit5.guice.module;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
 import com.google.inject.Provides;
 import io.qase.api.config.apiclient.ApiClientConfigurer;
-import io.qase.cucumber3.configuration.Cucumber3ApiConfigurer;
 import io.qase.guice.Injectors;
+import io.qase.junit5.configuration.Junit5ApiConfigurer;
 import lombok.Getter;
 
-public class Cucumber3Module extends AbstractModule {
+public class Junit5Module extends AbstractModule {
 
     @Getter
     private static final Injector injector = initializeInjector();
 
     @Provides
     public ApiClientConfigurer apiClientConfigurer() {
-        return new Cucumber3ApiConfigurer();
+        return new Junit5ApiConfigurer();
     }
 
     private static Injector initializeInjector() {
-        return Injectors.createDefaultInjectorWithOverridingModules(new Cucumber3Module());
+        return Injectors.createDefaultInjectorWithOverridingModules(new Junit5Module());
     }
 }
